owfd: p2pd: add wpa_supplicant interface
Add a new interface to spawn and connect to wpa_supplicant. Waiting for
wpa_supplicant to start up is kind of nasty as it doesn't support passing
control-sockets through on execve(). Hence, we need some magic to wait for
the socket to be ready via inotify.

Other than that, the interface is quite basic. More functions will follow
once we need them.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>