Import gdhcp from connman
Wifi-P2P needs dhcp to exchange addresses. As there is no decent dhcp
daemon out there, which can be controlled from an application ad-hoc, we
import gdhcp from connman.

gdhcp is a client and server implementation of dhcp controllable via a
simple API. We can use it to implement any kind of dhcp client and daemon
we want.

Unfortunately, gdhcp is GPLv2. That means all linking binaries are also
GPLv2. We don't care, but we need to explicitly mention it so others will
notice.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>