owfd: p2pd: ignore inotify_add_watch() failure
If the parent directory does not exist, yet, we cannot watch it. We would
have to watch the parent of it until the directory itself is created
(maybe multiple layers). Avoid this complexity by just polling with
reasonable timeouts instead.

Same for the file itself which might already exist (due to unclean
shutdown) and might get recreated.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>